backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
websockets
requests
werkzeug
alembic
httpx
